extract Guidance to it's own component so MenuList retains react-select innerRef
